Tried to implement wave
Tried to finish the Wave and WaveOne class. The wave works for up to 8
squirrels, but there is a bug where when there is 8+ squirrels some go
off the path. It also crashed when there is 12 or more. Will fix soon.

Created a new ScreenInfo class that the other classes will access to get
info about the screen and tile size.

Changed the parameter of PathFinder and Wave from map and map layer
number to the map layer itself because this was redundant.