Set new tmx file as background
I combined the final map image into a new tmx file that has the trees
blocked. Be warned that it is hard to make a maze at the moment because
it is not obvious which spots are blocked. If you get a lot of "path is
null" this is not an error. It will become easier when we implement
buying towers and not letting them block the path completely.