Added Resources and New Test Squirrel
To test out the SelectBox, I made a new Tower (called MaxTower, with
different visuals). I've also implemented resources into the game so you
cannot place a tower when you do not have the resources.

Creating new towers and adding different towers to the game should now
be easier. All you have to do is add a String into towerList in
GameScreen.java and change a bit of spawnTower(). See the bottom of
spawnTower().