Added the Labels
The Labels display the player's current life, wood, and stone.

Also, if a null path is found or the player lacks resources, an error
message will be displayed above the Label for displaying the amount of
wood the player has.

TO DO:

Add functionality to Resource and Life towers.
Balance the game.