Added Towers and Fixed Error Messages
Added Buff Towers -- Towers that Buff the damage of Stick Towers. So,
the base multiplier for damage is 1. For each Buff Tower, the damage
increases by +2 and is then halved. There's some math behind this but
basically, as you have more and more buff towers, the multiplier
increases slower.

Added functionality to Life Towers and Resource Towers.

Fixed errorMessage so that it auto-centers every time it gets updated.

To do: Balancing