Changed Projectile Picture
It bothered me that it was still just a giant red pixel thing so I
decided to just use some of my paint skills to make it look like a rock
(it's background is transparent so don't worry about that). It's
probably nothing compared to what Bryan can do but I think it's a nice
final polish!